+++ CHANGELOG.md
@@ -105,6 +105,7 @@ This update introduces some new tables and lots of changes.
 7. Fix Build error for PHP7 due to getID3 package. Switch to JamesHeinrich/getID3. (issue #152)
 8. Added delete image functionality to Bundles. (issue #149)
 9. Fixed Typeahead styling messed up for email suggestions. (issue #156)
+10. Upload image with same name will now be appended with number. (issue #154)
 
 ## Version 0.3.2 and 0.2.2
 Code clean up, new features and UI improvements.